Revert "Merge pull request #25 from AllenNeuralDynamics/test-upload-to-s3"
This reverts commit a28deb10f28c9205eb166ec748fff95c27ad33a9, reversing
changes made to cf71aa2651858b86a0483e1938905f8877d6f211.
