+++ src/aind_hcr_data_transformation/zeiss_job.py
@@ -1,17 +1,19 @@
 """Module to handle zeiss data compression"""
 
 import logging
+import asyncio
 import os
 import sys
 from pathlib import Path
 from time import time
-from typing import Any, Dict, List, Optional
+from typing import Any, Dict, List
 from urllib.parse import urlparse
+import multiprocessing
 
 from aind_data_transformation.core import GenericEtl, JobResponse, get_parser
 from packaging import version
 
-from aind_hcr_data_transformation.compress.czi_to_zarr_s3 import (
+from aind_hcr_data_transformation.compress.czi_to_zarr import (
     czi_stack_zarr_writer,
 )
 from aind_hcr_data_transformation.models import (
@@ -128,7 +130,7 @@ class ZeissCompressionJob(GenericEtl[ZeissJobSettings]):
 
         return [z, y, x]
 
-    def _get_compressor(self) -> Optional[Dict]:
+    def _get_compressor(self) -> Dict:
         """
         Utility method to construct a compressor class.
         Returns
@@ -194,9 +196,9 @@ class ZeissCompressionJob(GenericEtl[ZeissJobSettings]):
             )
             logging.info(msg)
 
-            czi_stack_zarr_writer(
+            asyncio.run(czi_stack_zarr_writer(
                 czi_path=str(stack),
-                output_path=str(output_path),
+                output_path=output_path,
                 voxel_size=voxel_size_zyx,
                 shard_size=self.job_settings.shard_size,
                 chunk_size=self.job_settings.chunk_size,
@@ -205,13 +207,16 @@ class ZeissCompressionJob(GenericEtl[ZeissJobSettings]):
                 downsample_mode=self.job_settings.downsample_mode,
                 channel_name=stack_name,
                 stack_name=f"{stack_name}.ome.zarr",
+                logger=logging,
                 compressor_kwargs=compressor,
                 bucket_name=bucket_name,
+                batch_size=self.job_settings.tensorstore_batch_size,
+            )
             )
 
     def _upload_derivatives_folder(self):
         """
-        Uploads the 'derivatives' folder inside of
+        Uploads the derivatives folder inside of
         the SPIM folder in the cloud.
         """
         s3_derivatives_dir = f"{self.job_settings.s3_location}/derivatives"
@@ -251,6 +256,7 @@ class ZeissCompressionJob(GenericEtl[ZeissJobSettings]):
         )
 
 
+# TODO: Add this to core aind_data_transformation class
 def job_entrypoint(sys_args: list):
     """Main function"""
     parser = get_parser()
@@ -270,4 +276,6 @@ def job_entrypoint(sys_args: list):
 
 
 if __name__ == "__main__":
+    multiprocessing.set_start_method("spawn", force=True)
     job_entrypoint(sys.argv[1:])
+
